Add 'system' static user as recognizable to the user service
This affects results in the following API endpoints:
* GET /user/{identifier}
* GET /information/username/{username}
* POST /users

It does not affect behaviour of
  GET /information/username/search/{username}
as that call is used to browse available users e.g. for assigning to
groups or to deck edit rights, so we don't want static users to be
listed there.

It also affects behaviour of POST /register, as it double checks for
already taken usernames in static users as well.